Updating PyThreadingPool and the PyMultiProcessTools package to be more 'pythonic' as well as give it some more features to excuse it for existing in the first place.
