--- PyThreadingPool/__init__.py
@@ -1,2 +1,0 @@
-import inspect, os, sys
-sys.path.append(os.path.dirname(os.path.abspath(inspect.getfile(inspect.currentframe()))))
